+++ IB-Tracker/src/Transactions/TransDetail.cpp
@@ -247,7 +247,7 @@ LRESULT CALLBACK TransDetail_ListBox_SubclassProc(
         // Accumulate delta until scroll one line (up +120, down -120). 
         // 120 is the Microsoft default delta
         int zDelta = GET_WHEEL_DELTA_WPARAM(wParam);
-        int top_index = SendMessage(hWnd, LB_GETTOPINDEX, 0, 0);
+        int top_index = (int)SendMessage(hWnd, LB_GETTOPINDEX, 0, 0);
         accumDelta += zDelta;
         if (accumDelta >= 120) {     // scroll up 3 lines
             top_index -= 3;
@@ -281,7 +281,7 @@ LRESULT CALLBACK TransDetail_ListBox_SubclassProc(
         SendMessage(hWnd, LB_GETITEMRECT, 0, (LPARAM)&rcItem);
         int itemHeight = (rcItem.bottom - rcItem.top);
         int items_count = ListBox_GetCount(hWnd);
-        int top_index = SendMessage(hWnd, LB_GETTOPINDEX, 0, 0);
+        int top_index = (int)SendMessage(hWnd, LB_GETTOPINDEX, 0, 0);
         int visible_rows = 0;
         int items_per_page = 0;
         int bottom_index = 0;
